Update BAG.cmd
Added overwrite and duplicate protection. Duplicate filenames dropped into the bag will be prepended with (n) when extracted to prevent them from overwriting eachother during extraction, leaving only the last copy dropped in. Alternatively, regardless of duplicates contained in the BAG, if a file already exists with the same name as the file to be extracted, the extracted filename will be prepended with (n) to its original name to avoid dataloss.

n refers to the next available number starting with 1